requirements.txt
asgiref==3.7.2
Brotli==1.0.9
dj-database-url==2.0.0
Django==3.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
whitenoise==6.5.0
